+++ provision/0020-vscode.sh
@@ -7,7 +7,10 @@ if [ -z "$provUser" ]; then
     exit 1
 fi
 
-curl -L https://go.microsoft.com/fwlink/?LinkID=760868 -o /tmp/vscode.deb 2>/dev/null
+if ! curl -L https://go.microsoft.com/fwlink/?LinkID=760868 -o /tmp/vscode.deb 2>/dev/null; then
+    exit 1
+fi
+
 sudo apt-get update
 sudo apt-get install -y -f /tmp/vscode.deb libxss1
 rm -f /tmp/vscode.deb
